Conflicts:
	mobile/android/SyncProxyAndroid/src/main/java/com/ford/syncV4/protocol/ProtocolMessage.java
	mobile/android/SyncProxyAndroid/src/main/java/com/ford/syncV4/protocol/ProtocolMessageConverter.java
	mobile/android/SyncProxyAndroid/src/main/java/com/ford/syncV4/proxy/RPCRequest.java
	mobile/android/SyncProxyAndroid/src/main/java/com/ford/syncV4/proxy/SyncProxyBase.java
	mobile/android/SyncProxyTester/src/main/AndroidManifest.xml
	mobile/android/SyncProxyTester/src/main/assets/CHANGELOG.txt

+++ src/components/application_manager/include/application_manager/application_manager_impl.h
@@ -554,7 +554,7 @@ class ApplicationManagerImpl : public ApplicationManager,
         const std::vector<uint8_t>& binary_data,
         const std::string& file_path,
         const std::string& file_name,
-        const uint32_t offset);
+        const int64_t offset);
 
     /**
      * @brief Get available app space
